requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.1
whitenoise==6.7.0
selenium